Squeeze warnings on msvc express 2010.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@36344 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/Paragraph.cpp
@@ -1111,7 +1111,7 @@ void Paragraph::Private::latexInset(BufferParams const & bparams,
 		texrow.start(owner_->id(), i + 1);
 		column = 0;
 	} else {
-		column += os.tellp() - len;
+		column += (unsigned int)(os.tellp() - len);
 	}
 
 	if (owner_->isDeleted(i))
